Fix Comics Manager plugin type errors

+++ plugins/python/comics_project_management_tools/comics_project_manager_docker.py
@@ -136,8 +136,8 @@ class comic_page_delegate(QStyledItemDelegate):
             
             descRect = QRect(textRect.left(), textRect.bottom()+margin, labelWidth, (rect.bottom()-margin) - (textRect.bottom()+margin))
             if textRect.bottom()+metrics.height() < rect.bottom():
-                textRect.setBottom(textRect.bottom()+(margin/2))
-                textRect.setLeft(textRect.left()-(margin/2))
+                textRect.setBottom(int(textRect.bottom()+(margin/2)))
+                textRect.setLeft(int(textRect.left()-(margin/2)))
                 painter.setOpacity(0.4)
                 painter.drawLine(textRect.bottomLeft(), textRect.bottomRight())
                 painter.setOpacity(1.0)
